API

IntervalRootFinding.bisect — Function.

bisect(X::IntervalBox, α=0.5)

Bisect the IntervalBox X at position α ∈ [0,1] along its longest side.

source

IntervalRootFinding.bisect — Function.

bisect(X::Interval, α=0.5)

Split the interval X at position α; α=0.5 corresponds to the midpoint. Returns a tuple of the new intervals.

source

IntervalRootFinding.bisect — Function.

bisect(X::IntervalBox, i::Integer, α=0.5)

Bisect the IntervalBox in side number i.

source

IntervalRootFinding.bisection — Method.

bisection(f, X; tolerance=1e-3)

Find possible roots of the function f inside the Interval or IntervalBox X.

source

IntervalRootFinding.newton — Function.

newton performs the interval Newton method on the given function f with its optional derivative f_prime and initial interval x. Optional keyword arguments give the tolerance, maxlevel at which to stop subdividing, and a debug boolean argument that prints out diagnostic information.

source

IntervalRootFinding.guarded_derivative_midpoint — Method.

Returns two intervals, the first being a point within the interval x such that the interval corresponding to the derivative of f there does not contain zero, and the second is the inverse of its derivative

source

IntervalRootFinding.guarded_mid — Method.

Returns the midpoint of the interval x, slightly shifted in case the midpoint is an exact root

source

IntervalRootFinding.guarded_mid — Method.

Returns the midpoint of the interval x, slightly shifted in case the midpoint is an exact root

source

IntervalRootFinding.newton_refine — Method.

If a root is known to be inside an interval, newton_refine iterates the interval Newton method until that root is found.

source