API

Main interface

IntervalRootFinding.branch_and_prune — Method

branch_and_prune(X, contractor, search_order, tol)

Generic branch and prune routine for finding isolated roots using the given contractor to determine the status of a given box X.

See the documentation of the roots function for explanation of the other arguments.

Polynomials.roots — Method

roots(f, X, contractor=Newton, search_order=BreadthFirst, tol=1e-15)
roots(f, deriv, X, contractor=Newton, search_order=BreadthFirst, tol=1e-15)
roots(f, X, contractor, tol)
roots(f, deriv, X, contractor, tol)

Uses a generic branch and prune routine to find in principle all isolated roots of a function f:R^n → R^n in a region X, if the number of roots is finite.

Inputs:

f: function whose roots will be found

X: Interval or IntervalBox in which roots are searched

contractor: function that, when applied to the function f, determines the status of a given box X. It returns the new box and a symbol indicating the status. Current possible values are Bisection, Newton and Krawczyk

deriv: explicit derivative of f for Newton and Krawczyk

search_order: SearchStrategy determining the order in which regions are processed.

tol: Absolute tolerance. If a region has a diameter smaller than tol, it is returned with status :unknown.

Root object

IntervalRootFinding.Root — Type

Root

Object representing a possible root inside a given region. The field status is either :unknown or :unique. If status is :unique then we know that there is a unique root of the function in question inside the given region.

Internally the status may also be :empty for region guaranteed to contain no root, however such Roots are discarded by default and thus never returned by the roots function.

Fields

interval: a region (either Interval of IntervalBox) searched for roots.

status: the status of the region, valid values are :empty, unknown and :unique.

IntervalRootFinding.isunique — Method

isunique(rt)

Return whether a Root is unique.

IntervalRootFinding.root_region — Method

root_region(rt)

Return the region associated to a Root.

IntervalRootFinding.root_status — Method

root_status(rt)

Return the status of a Root.

Contractors

IntervalRootFinding.AbstractContractor — Type

AbstractContractor{F}

Abstract type for contractors.

IntervalRootFinding.Bisection — Type

Bisection{F} <: AbstractContractor{F}

AbstractContractor type for the bisection method.

IntervalRootFinding.Krawczyk — Type

Krawczyk{F, FP} <: AbstractContractor{F}

AbstractContractor type for the interval Krawczyk method.

Fields

- `f::F`: function whose roots are searched
- `f::FP`: derivative or jacobian of `f`

(K::Krawczyk)(X ; α=where_bisect)

Contract an interval X using Krawczyk operator and return the contracted interval together with its status.

Inputs

- `R`: Root object containing the interval to contract.
- `α`: Point of bisection of intervals.

IntervalRootFinding.Newton — Type

Newton{F, FP} <: AbstractContractor{F}

AbstractContractor type for the interval Newton method.

Fields

- `f::F`: function whose roots are searched
- `f::FP`: derivative or jacobian of `f`

(N::Newton)(X, α=where_bisect)

Contract an interval X using Newton operator and return the contracted interval together with its status.

Inputs

- `R`: Root object containing the interval to contract.
- `α`: Point of bisection of intervals.

IntervalRootFinding.contract — Method

contract(contractor, R)

Contract the region R using the given contractor.

IntervalRootFinding.refine — Method

refine(op, X::Tuple{Symbol, Region}, tol)

Wrap the refine method to leave unchanged intervals that are not guaranteed to contain an unique solution.

IntervalRootFinding.refine — Method

refine(C, X::Region, tol)

Refine a interval known to contain a solution.

This function assumes that it is already known that X contains a unique root.

IntervalRootFinding.safe_isempty — Method

safe_isempty(X)

Similar to isempty function for IntervalBox, but also works for SVector of Interval.

Branch-and-bound search interface

Others

IntervalRootFinding.Compl — Type

Complex numbers as 2-vectors, enough for polynomials.

IntervalRootFinding.realify — Method

Takes a complex (polynomial) function f and returns a function g:R^2 -> R^2 that implements it.

IntervalRootFinding.realify_derivative — Method

Takes the derivative of a complex function and returns the real jacobian that implements it.

IntervalRootFinding.gauss_elimination_interval! — Method

Solves the system of linear equations using Gaussian Elimination. Preconditioning is used when the precondition keyword argument is true.

REF: Luc Jaulin et al., Applied Interval Analysis, pg. 72

IntervalRootFinding.gauss_elimination_interval1! — Method

Using Base.`

IntervalRootFinding.gauss_seidel_interval! — Method

Iteratively solves the system of interval linear equations and returns the solution set. Uses the Gauss-Seidel method (Hansen-Sengupta version) to solve the system. Keyword precondition to turn preconditioning off. Eldon Hansen and G. William Walster : Global Optimization Using Interval Analysis - Chapter 5 - Page 115

IntervalRootFinding.preconditioner — Method

Preconditions the matrix A and b with the inverse of mid(A)

IntervalRootFinding.newton1d — Method

newton1d performs the interval Newton method on the given function f with its derivative f′ and initial interval x. Optional keyword arguments give the tolerances reltol and abstol. reltol is the tolerance on the relative error whereas abstol is the tolerance on |f(X)|, and a debug boolean argument that prints out diagnostic information.

IntervalRootFinding.newton1d — Method

newton1d performs the interval Newton method on the given function f and initial interval x. Optional keyword arguments give the tolerances reltol and abstol. reltol is the tolerance on the relative error whereas abstol is the tolerance on |f(X)|, and a debug boolean argument that prints out diagnostic information.

IntervalRootFinding.quadratic_helper! — Method

Helper function for quadratic_interval that computes roots of a real quadratic using interval arithmetic to bound rounding errors.

IntervalRootFinding.quadratic_roots — Method

Function to solve a quadratic equation where the coefficients are intervals. Returns an array of intervals of the roots. Arguments a, b and c are interval coefficients of x², x and 1 respectively. The interval case differs from the non-interval case in that there might be three disjoint interval roots. In the third case, one interval root extends to −∞ and another extends to +∞. This algorithm finds the set of points where F.lo(x) ≥ 0 and the set of points where F.hi(x) ≤ 0 and takes the intersection of these two sets. Eldon Hansen and G. William Walster : Global Optimization Using Interval Analysis - Chapter 8